backend/requirements.txt
fastapi
sqlalchemy
pydantic
psycopg2-binary
python-jose
passlib[bcrypt]
